+++ src/nas.cpp
@@ -2,7 +2,7 @@
 #include <cpp11.hpp>
 #include <Rinternals.h>
 
-R_xlen_t na_count(SEXP x){
+R_xlen_t na_count(SEXP x, bool recursive){
   R_xlen_t n = Rf_xlength(x);
   R_xlen_t count = 0;
   int n_protections = 0;
@@ -79,14 +79,21 @@ R_xlen_t na_count(SEXP x){
     // ++n_protections;
     // int *p_is_empty = LOGICAL(is_empty);
     // count = count_true(p_is_empty, num_row);
+    // if (recursive){
     const SEXP *p_x = VECTOR_PTR_RO(x);
     for (R_xlen_t i = 0; i < n; ++i){
-      count += na_count(p_x[i]);
+      count += na_count(p_x[i], true);
     }
-    break;
+  // } else {
+  if (recursive) break;
+  // }
+    // break;
   }
   default: {
-    Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(x)));
+    SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](x));
+    ++n_protections;
+    int *p_is_missing = LOGICAL(is_missing);
+    count = count_true(p_is_missing, Rf_xlength(is_missing));
     break;
   }
   }
@@ -95,12 +102,13 @@ R_xlen_t na_count(SEXP x){
 }
 
 [[cpp11::register]]
-SEXP cpp_num_na(SEXP x){
-  return xlen_to_r(na_count(x));
+SEXP cpp_num_na(SEXP x, bool recursive){
+  return xlen_to_r(na_count(x, recursive));
 }
 
 [[cpp11::register]]
-bool cpp_any_na(SEXP x){
+bool cpp_any_na(SEXP x, bool recursive){
+  int n_protections = 0;
   R_xlen_t n = Rf_xlength(x);
   bool out = false;
   if (n == 0){
@@ -113,7 +121,7 @@ bool cpp_any_na(SEXP x){
     // do a pseudo while-loop in openMP
     // volatile bool flag = false;
     int *p_x = INTEGER(x);
-// #pragma omp parallel for num_threads(num_cores()) shared(flag)
+    // #pragma omp parallel for num_threads(num_cores()) shared(flag)
     for (R_xlen_t i = 0; i < n; ++i){
       // if (flag) continue;
       if (p_x[i] == NA_INTEGER){
@@ -159,21 +167,27 @@ bool cpp_any_na(SEXP x){
   }
   case VECSXP: {
     for (int i = 0; i < n; ++i){
-    out = cpp_any_na(VECTOR_ELT(x, i));
-    if (out) break;
-  }
-    break;
+      out = cpp_any_na(VECTOR_ELT(x, i), true);
+      if (out) break;
+    }
+    if (recursive) break;
   }
   default: {
-    Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(x)));
+    SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](x));
+    ++n_protections;
+    SEXP any_missing = Rf_protect(cpp11::package("base")["any"](is_missing));
+    ++n_protections;
+    out = Rf_asLogical(any_missing);
     break;
   }
   }
+  Rf_unprotect(n_protections);
   return out;
 }
 
 [[cpp11::register]]
-bool cpp_all_na(SEXP x, bool return_true_on_empty){
+bool cpp_all_na(SEXP x, bool return_true_on_empty, bool recursive){
+  int n_protections = 0;
   R_xlen_t n = Rf_xlength(x);
   bool out = true;
   if (n == 0){
@@ -230,16 +244,21 @@ bool cpp_all_na(SEXP x, bool return_true_on_empty){
   }
   case VECSXP: {
     for (int i = 0; i < n; ++i){
-    out = cpp_all_na(VECTOR_ELT(x, i), return_true_on_empty);
-    if (!out) break;
-  }
-    break;
+      out = cpp_all_na(VECTOR_ELT(x, i), return_true_on_empty, true);
+      if (!out) break;
+    }
+    if (recursive) break;
   }
   default: {
-    Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(x)));
+    SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](x));
+    ++n_protections;
+    SEXP all_missing = Rf_protect(cpp11::package("base")["all"](is_missing));
+    ++n_protections;
+    out = Rf_asLogical(all_missing);
     break;
   }
   }
+  Rf_unprotect(n_protections);
   return out;
 }
 
@@ -257,7 +276,7 @@ SEXP cpp_which_na(SEXP x){
   switch ( TYPEOF(x) ){
   case LGLSXP:
   case INTSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     int *p_x = INTEGER(x);
     if (is_short){
       SEXP out = Rf_protect(Rf_allocVector(INTSXP, count));
@@ -286,7 +305,7 @@ SEXP cpp_which_na(SEXP x){
     }
   }
   case REALSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     double *p_x = REAL(x);
     if (is_short){
       SEXP out = Rf_protect(Rf_allocVector(INTSXP, count));
@@ -315,8 +334,7 @@ SEXP cpp_which_na(SEXP x){
     }
   }
   case STRSXP: {
-    R_xlen_t count = na_count(x);
-    // R_xlen_t count = Rf_asReal(Rf_protect(cpp_num_na(x)));
+    R_xlen_t count = na_count(x, true);
     SEXP *p_x = STRING_PTR(x);
     if (is_short){
       SEXP out = Rf_protect(Rf_allocVector(INTSXP, count));
@@ -350,7 +368,7 @@ SEXP cpp_which_na(SEXP x){
     return out;
   }
   case CPLXSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     Rcomplex *p_x = COMPLEX(x);
     if (is_short){
       SEXP out = Rf_protect(Rf_allocVector(INTSXP, count));
@@ -378,17 +396,11 @@ SEXP cpp_which_na(SEXP x){
       return out;
     }
   }
-  // case VECSXP: {
-  //   if (Rf_inherits(x, "vctrs_rcrd")){
-  //   SEXP rcrd_df = Rf_protect(cpp11::package("cheapr")["vctrs_rcrd_as_df"](x));
-  //   SEXP is_empty = Rf_protect(cpp_missing_row(rcrd_df, 1, true));
-  //   SEXP out = Rf_protect(cpp_which_(is_empty, false));
-  //   Rf_unprotect(3);
-  //   return out;
-  // }
-  // }
   default: {
-    Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(x)));
+    SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](x));
+    SEXP out = Rf_protect(cpp_which_(is_missing, false));
+    Rf_unprotect(2);
+    return out;
     break;
   }
   }
@@ -406,7 +418,7 @@ SEXP cpp_which_not_na(SEXP x){
   switch ( TYPEOF(x) ){
   case LGLSXP:
   case INTSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     int *p_x = INTEGER(x);
     if (is_short){
       int out_size = n - count;
@@ -437,7 +449,7 @@ SEXP cpp_which_not_na(SEXP x){
     }
   }
   case REALSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     double *p_x = REAL(x);
     if (is_short){
       int out_size = n - count;
@@ -468,7 +480,7 @@ SEXP cpp_which_not_na(SEXP x){
     }
   }
   case STRSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     SEXP *p_x = STRING_PTR(x);
     if (is_short){
       int out_size = n - count;
@@ -518,7 +530,7 @@ SEXP cpp_which_not_na(SEXP x){
   }
   }
   case CPLXSXP: {
-    R_xlen_t count = na_count(x);
+    R_xlen_t count = na_count(x, true);
     Rcomplex *p_x = COMPLEX(x);
     if (is_short){
       int out_size = n - count;
@@ -548,17 +560,11 @@ SEXP cpp_which_not_na(SEXP x){
       return out;
     }
   }
-    // case VECSXP: {
-    //   if (Rf_inherits(x, "vctrs_rcrd")){
-    //   SEXP rcrd_df = Rf_protect(cpp11::package("cheapr")["vctrs_rcrd_as_df"](x));
-    //   SEXP is_empty = Rf_protect(cpp_missing_row(rcrd_df, 1, true));
-    //   SEXP out = Rf_protect(cpp_which_(is_empty, true));
-    //   Rf_unprotect(3);
-    //   return out;
-    // }
-    // }
   default: {
-    Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(x)));
+    SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](x));
+    SEXP out = Rf_protect(cpp_which_(is_missing, true));
+    Rf_unprotect(2);
+    return out;
     break;
   }
   }
@@ -657,43 +663,23 @@ SEXP cpp_is_na(SEXP x){
     // break;
   }
   case VECSXP: {
-  //   if (Rf_inherits(x, "vctrs_rcrd")){
-  //   SEXP rcrd_df = Rf_protect(cpp11::package("cheapr")["vctrs_rcrd_as_df"](x));
-  //   SEXP out = Rf_protect(cpp_missing_row(rcrd_df, 1, true));
-  //   Rf_unprotect(2);
-  //   return out;
-  // } else {
   if (!Rf_isObject(x)){
     SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
     int *p_out = LOGICAL(out);
     const SEXP *p_x = VECTOR_PTR_RO(x);
     for (i = 0; i < n; ++i){
-      p_out[i] = cpp_all_na(p_x[i], false);
+      p_out[i] = cpp_all_na(p_x[i], false, true);
     }
     Rf_unprotect(1);
     return out;
   }
-  // SEXP out = Rf_protect(Rf_allocVector(LGLSXP, n));
-  //   int *p_out = LOGICAL(out);
-  //   const SEXP *p_x = VECTOR_PTR_RO(x);
-  //   for (i = 0; i < n; ++i){
-  //     p_out[i] = cpp_all_na(p_x[i], false);
-  //   }
-  //   Rf_unprotect(1);
-  //   return out;
-  // }
-    // break;
   }
   default: {
     SEXP out = Rf_protect(cpp11::package("base")["is.na"](x));
     Rf_unprotect(1);
     return out;
-    // Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(x)));
-    // break;
   }
   }
-  // Rf_unprotect(n_protections);
-  // return out;
 }
 
 [[cpp11::register]]
@@ -762,40 +748,57 @@ SEXP cpp_row_na_counts(SEXP x){
       break;
     }
     case VECSXP: {
-      // SEXP is_empty_nested = Rf_protect(cpp_missing_row(VECTOR_ELT(x, j), 1, true));
-      // ++n_protections;
-      // int *p_is_empty_nested = LOGICAL(is_empty_nested);
-      // for (R_xlen_t k = 0; k < num_row; ++k){
-      //   p_n_empty[k] += p_is_empty_nested[k];
-      // }
-
-
-      // if (Rf_inherits(p_x[j], "vctrs_rcrd")){
-      //   SEXP rcrd_df = Rf_protect(cpp11::package("cheapr")["vctrs_rcrd_as_df"](p_x[j]));
-      //   ++n_protections;
-      //   SEXP is_empty_nested = Rf_protect(cpp_missing_row(rcrd_df, 1, true));
-      //   ++n_protections;
-      //   int *p_is_empty_nested = LOGICAL(is_empty_nested);
-      //   for (R_xlen_t k = 0; k < num_row; ++k){
-      //     p_n_empty[k] += p_is_empty_nested[k];
-      //   }
-      // } else {
-      if (Rf_xlength(p_x[j]) != num_row){
-      int int_nrows = num_row;
-      int element_length = Rf_xlength(p_x[j]);
+      if (Rf_isObject(p_x[j])){
+      SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](p_x[j]));
       ++n_protections;
-      SEXP names = Rf_protect(Rf_getAttrib(x, R_NamesSymbol));
-      Rf_unprotect(n_protections);
-      Rf_error("list variable %s has length (%d) not equal to number of rows (%d)",
-               CHAR(STRING_ELT(names, j)), element_length, int_nrows);
-    }
+      if (Rf_xlength(is_missing) != num_row){
+        int int_nrows = num_row;
+        int element_length = Rf_xlength(is_missing);
+        ++n_protections;
+        SEXP names = Rf_protect(Rf_getAttrib(x, R_NamesSymbol));
+        Rf_unprotect(n_protections);
+        Rf_error("is.na method for list variable %s produces a length (%d) vector which does not equal the number of rows (%d)",
+                 CHAR(STRING_ELT(names, j)), element_length, int_nrows);
+      }
+      int *p_is_missing = LOGICAL(is_missing);
+      for (R_xlen_t k = 0; k < num_row; ++k){
+        p_n_empty[k] += p_is_missing[k];
+      }
+    } else {
       const SEXP *p_xj = VECTOR_PTR_RO(p_x[j]);
       for (R_xlen_t i = 0; i < num_row; ++i){
-        p_n_empty[i] += cpp_all_na(p_xj[i], false);
+        p_n_empty[i] += cpp_all_na(p_xj[i], false, true);
       }
-      // }
-      break;
     }
+    break;
+    }
+    // case VECSXP: {
+    //   if (cpp_is_list_df_like(p_x[j])){
+    //     SEXP df = Rf_protect(cpp11::package("cheapr")["list_as_df"](p_x[j]));
+    //     ++n_protections;
+    //     SEXP is_empty_nested = Rf_protect(cpp_missing_row(df, 1, true));
+    //     ++n_protections;
+    //     int *p_is_empty_nested = LOGICAL(is_empty_nested);
+    //     for (R_xlen_t k = 0; k < num_row; ++k){
+    //       p_n_empty[k] += p_is_empty_nested[k];
+    //     }
+    //   } else if (Rf_xlength(p_x[j]) != num_row){
+    //     // Throw an error here
+    //     int int_nrows = num_row;
+    //     int element_length = Rf_xlength(p_x[j]);
+    //     ++n_protections;
+    //     SEXP names = Rf_protect(Rf_getAttrib(x, R_NamesSymbol));
+    //     Rf_unprotect(n_protections);
+    //     Rf_error("list variable %s has length (%d) not equal to number of rows (%d)",
+    //              CHAR(STRING_ELT(names, j)), element_length, int_nrows);
+    //   } else {
+    //   const SEXP *p_xj = VECTOR_PTR_RO(p_x[j]);
+    //   for (R_xlen_t i = 0; i < num_row; ++i){
+    //     p_n_empty[i] += cpp_all_na(p_xj[i], false);
+    //   }
+    // }
+    //   break;
+    // }
     default: {
       Rf_unprotect(n_protections);
       Rf_error("%s cannot handle an object of type %s", __func__, Rf_type2char(TYPEOF(p_x[j])));
@@ -822,33 +825,56 @@ SEXP cpp_col_na_counts(SEXP x){
   for (int j = 0; j < num_col; ++j){
     switch ( TYPEOF(p_x[j]) ){
     case VECSXP: {
-      //   if (Rf_inherits(p_x[j], "vctrs_rcrd")){
-      //   SEXP rcrd_df = Rf_protect(cpp11::package("cheapr")["vctrs_rcrd_as_df"](p_x[j]));
-      //   ++n_protections;
-      //   SEXP is_empty_nested = Rf_protect(cpp_missing_row(rcrd_df, 1, true));
-      //   ++n_protections;
-      //   int *p_is_empty_nested = LOGICAL(is_empty_nested);
-      //   for (R_xlen_t k = 0; k < num_row; ++k){
-      //     p_out[j] += p_is_empty_nested[k];
-      //   }
-      // } else {
-      if (Rf_xlength(p_x[j]) != num_row){
-      int int_nrows = num_row;
-      int element_length = Rf_xlength(p_x[j]);
+      if (Rf_isObject(p_x[j])){
+      SEXP is_missing = Rf_protect(cpp11::package("cheapr")["is_na"](p_x[j]));
       ++n_protections;
-      SEXP names = Rf_protect(Rf_getAttrib(x, R_NamesSymbol));
-      Rf_unprotect(n_protections);
-      Rf_error("list variable %s has length (%d) not equal to number of rows (%d)",
-               CHAR(STRING_ELT(names, j)), element_length, int_nrows);
-    }
-      for (R_xlen_t i = 0; i < num_row; ++i){
-        p_out[j] += cpp_all_na(VECTOR_ELT(p_x[j], i), false);
+      if (Rf_xlength(is_missing) != num_row){
+        int int_nrows = num_row;
+        int element_length = Rf_xlength(is_missing);
+        ++n_protections;
+        SEXP names = Rf_protect(Rf_getAttrib(x, R_NamesSymbol));
+        Rf_unprotect(n_protections);
+        Rf_error("is.na method for list variable %s produces a length (%d) vector which does not equal the number of rows (%d)",
+                 CHAR(STRING_ELT(names, j)), element_length, int_nrows);
       }
-      // }
-      break;
+      int *p_is_missing = LOGICAL(is_missing);
+      for (R_xlen_t k = 0; k < num_row; ++k){
+        p_out[j] += p_is_missing[k];
+      }
+    } else {
+      for (R_xlen_t i = 0; i < num_row; ++i){
+        p_out[j] += cpp_all_na(VECTOR_ELT(p_x[j], i), false, true);
+      }
     }
+    break;
+    }
+    // case VECSXP: {
+    //   if (cpp_is_list_df_like(p_x[j])){
+    //   SEXP df = Rf_protect(cpp11::package("cheapr")["list_as_df"](p_x[j]));
+    //   ++n_protections;
+    //   SEXP is_empty_nested = Rf_protect(cpp_missing_row(df, 1, true));
+    //   ++n_protections;
+    //   int *p_is_empty_nested = LOGICAL(is_empty_nested);
+    //   for (R_xlen_t k = 0; k < num_row; ++k){
+    //     p_out[j] += p_is_empty_nested[k];
+    //   }
+    // } else if (Rf_xlength(p_x[j]) != num_row){
+    //   int int_nrows = num_row;
+    //   int element_length = Rf_xlength(p_x[j]);
+    //   ++n_protections;
+    //   SEXP names = Rf_protect(Rf_getAttrib(x, R_NamesSymbol));
+    //   Rf_unprotect(n_protections);
+    //   Rf_error("list variable %s has length (%d) not equal to number of rows (%d)",
+    //            CHAR(STRING_ELT(names, j)), element_length, int_nrows);
+    // } else {
+    //   for (R_xlen_t i = 0; i < num_row; ++i){
+    //     p_out[j] += cpp_all_na(VECTOR_ELT(p_x[j], i), false);
+    //   }
+    // }
+    // break;
+    // }
     default: {
-      p_out[j] = na_count(p_x[j]);
+      p_out[j] = na_count(p_x[j], false);
       break;
     }
     }