+++ NEWS.md
@@ -1,4 +1,13 @@
-# cheapr 0.2.0
+# cheapr 0.3.0
+
+* New function `overview` as a cheaper alternative to `summary`.
+
+* All of the `NA` handling functions now fall back to using `is.na` if an appropriate
+method cannot be found.
+
+* More support has been added for all objects with an `is.na` method.
+
+# cheapr 0.2.0 (06-Mar-2024)
 
 * `is_na` has been added as an S3 generic function which is parallelised and  internally falls back
 on `is.na` if there are no suitable methods.